Move all files to pxf directory and preserve PXF-specific commits
We are doing this in preparation for bringing in the automation framework
tests from the (currently closed-source) infrastructure repo.

We followed the instructions in this post to select only the commits
from Apache Hawq that affect PXF code:

http://gbayer.com/development/moving-files-from-one-git-repository-to-another-preserving-history/

That post recommends using git filter-branch and then making a new
commit that moves the files to the desired directory. This is
that commit.

Co-authored-by: Shivram Mani <smani@pivotal.io>
Co-authored-by: Ben Christel <bchristel@pivotal.io>