Create a template for the release and PXF pipelines (#47)
The PXF and release pipelines are very similar. When changing one pipeline we had the risk of the other pipeline becoming out of sync. For that reason, we introduce a template with a single source of truth for both pipelines.
We introduce jinja2, following a similar approach from GPDB prod pipeline compilation. We also combine the logic in the existing PXF deploy script and the new pipeline generation scripts.

Templating the pipeline and integrating it to the deploy script also allows us to provide an optional flag to generate a pipeline with a compile job for GPDB.

Finally, we add the option to deploy an acceptance pipeline.

[#160429794] [#158565980]