pxf cluster: stop checking that hostname is master
Checking that the user is running `pxf cluster` from master is helpful
for ensuring that the source of truth for PXF configs lives on master.
We don't want users doing cluster configuration from segment-only hosts.

However, this causes issues when the hostname reported by the OS doesn't
match what Greenplum knows about, so we are foregoing this check for
now.

Authored-by: Oliver Albertini <oalbertini@pivotal.io>