pxf-cli: Support sync and init on standby master
This is a temporary solution until this issue on gp-common-go-libs is
fixed: https://github.com/greenplum-db/gp-common-go-libs/issues/38

This issue would allow us to simply add standby/mirrors to the hosts we
want to execute on. Instead, in this commit, we are using a direct query
to the database to find out standby master (contentID: -1, not primary).
With the standby, we add it to our list of segments early in execution,
but only if we are doing sync or init.

Authored-by: Oliver Albertini <oalbertini@pivotal.io>