Kill JVM/Tomcat on OutOfMemoryError
After an OutOfMemoryError, the JVM and Tomcat instance are considered
unreliable, and it's best to kill the process. In this commit, we
introduce a script `kill-pxf.sh` in the pxf-service/bin that handles
this situation.

A sleep is used in `kill-pxf.sh` to allow the JVM to log and bubble up
to the user the OutOfMemoryError.

For testing we have an OutOfMemoryFragmenter that creates a very large
list, running into the memory limits. This should bring the app down,
and we sleep for 3 seconds to give the app time to be brought down, then
check that we get Failed to connect error.

We also support the user not killing the JVM on OOM by setting the
environment variable PXF_OOM_KILL to false in `$PXF_CONF/conf/pxf-env.sh`.

The user can also set PXF_OOM_DUMP_PATH to a directory or a file path to
get a heap dump on OOM. If it's a directory, Tomcat will make a heap
dump file called `java_pid<PID>.hprof` at that directory.

We give the user a NOTICE when they set PXF_OOM_DUMP_PATH informing them
that they need enough disk space to hold a heap dump file there.

Co-authored-by: Oliver Albertini <oalbertini@pivotal.io>
Co-authored-by: Alex Denissov <adenissov@pivotal.io>
Co-authored-by: Francisco Guerrero <aguerrero@pivotal.io>
Co-authored-by: Shivram Mani <smani@pivotal.io>
Co-authored-by: Lav Jain <ljain@pivotal.io>