Fix paramiko pipeline failure
Python library Paramiko (which is required for testing with Tinc) was
recently removed from the python-dependencies.txt file of the GPDB repo
as well as the Dockerfiles in `gpdb/src/tools/docker`.  This led to
several pipeline failures when testing against GPDB master and
6X_STABLE. On 5X, paramiko is still being bundled inside of
`bin_gpdb.tar.gz`, so our 5X builds were unaffected.

Paramiko was available on the CentOS concourse boxes for a while because
we were not triggering our `cdh`, `hdp`, and `mapr` server builds. This
commit corrects that, and ensures that PYTHONPATH in the Tinc test
environment points to system python libraries for CentOS 6/7 and Ubuntu
16. This ensures that things in the system python libraries can be
accessed by Tinc, including paramiko, its dependencies, and another
python library `psi`.

The pxf-dev-base Dockerfile now installs paramiko. Some logic is
required to do this in the Dockerfile since CentOS 6 uses python 2.6 and
paramiko doesn't support python 2.6 anymore (`pip install` won't work in
that case, we rely on `yum`).

Finally, to support Ubuntu better, we switch to only strict POSIX shell
features in Makefiles. In Ubuntu, the POSIX shell is dash and it doesn't
support `[[` or `==` notations.

Co-authored-by: Oliver Albertini <oalbertini@pivotal.io>
Co-authored-by: Raymond Yin <ryin@pivotal.io>
Co-authored-by: Alex Denissov <adenissov@pivotal.io>