JsonResolver: throw BadRecordException on bad JSON
Previously, when a user created an external JSON table containing
records such as `{ "key": "(" }`, where the value was cast to a number
such as SMALLINT in the database, we sent along zero as the number.  If the
value was unquoted: `{ "key": ( }` (bad JSON formatting), we sent an
empty row.

This commit corrects that behavior by validating numeric types and
booleans. If malformatted JSON or an invalid type is found, it raises
BadRecordException which GpdbWritable can send to the database as an
error for the row. The database can then report the error, which
can be suppressed for unclean data using the REJECT LIMIT option. The
errors can be read by querying e.g. in GPDB like this:

select * from gp_read_error_log('table_name');

Also, we start using the more modern Jackson (FasterXML, not codehaus)
in JsonResolver.

We also updated the JsonExtensionTest to reflect these changes, and
added automation tests for malformatted JSON and mismatched types with
and without REJECT LIMIT and LOG ERROR set.

Adds unit tests for JsonResolver

Tests all numeric and boolean cases for mismatched types, and also
malformatted JSON.

Finally, we also start testing JSON with HCFS in automation.

Co-authored-by: Oliver Albertini <oalbertini@pivotal.io>
Co-authored-by: Alexander Denissov <adenissov@pivotal.io>
Co-authored-by: Francisco Guerrero <aguerrero@pivotal.io>
Co-authored-by: Raymond Yin <ryin@pivotal.io>