JDBC: Query data from ranges outside of partition range (#170)
Modify JDBC partitioning in order to make it a perfomance-increasing feature not affecting the SELECT query output.

In the current version of this PR, this, this and this comments are taken into account. They provide an explanation of the motivation for the changes.

The partition boundaries (in INT and DATE partitions) given in RANGE setting of LOCATION clause are now considered to form a closed interval (including both its lower and upper boundaries). Previously, the interval they formed was considered a left-closed one (not including its upper boundary).

Example
The following example illustrates the changes in API.

Consider the following string (a part of LOCATION clause):

...&PARTITION_BY=column:int&RANGE=1:5&INTERVAL=2&...

Old behaviour led to generation of fragments with constraints:

>= 1 AND < 3
>= 3 AND < 5
New behaviour leads to generation of fragments with constraints:

< 1
>= 5
>= 1 AND < 3
>= 3 AND < 5
IS NULL
Code
To implement the changes in API, a new (sub)package org.greenplum.pxf.plugins.jdbc.partitioning is introduced.

Each partition type is processed by enum PartitionType. The results are serialized using Apache SerializationUtils (this library is already used in PXF). This enables SQLQueryBuilder to use a common interface for all partition types and moves all partitioning logic to org.greenplum.pxf.plugins.jdbc.partitioning package and JdbcPartitionFragmenter class.

Other changes
Update README.md
Add unit tests for the new feature and update the existing ones