Allow dynamic Accessor/Resolver via the ProtocolHandler interface
We add the ProtocolHandler interface, that allows developers to choose
their own Accessor and Resolver based on the the RequestContext. We
implement the S3ProtocolHandler, which dynamically decides which
Accessor/Resolver (for S3 Select) to use.

We now support parsing the protocol and format correctly when
interacting with both the external table framework and the foreign data
wrapper framework. The WIRE FORMAT refers to the communication protocol
between Greenplum and PXF, and the FORMAT refers to the format of the
file in the external system (i.e CSV, Parquet, AVRO, etc).

Co-authored-by: Francisco Guerrero <aguerrero@pivotal.io>
Co-authored-by: Alexander Denissov <adenissov@pivotal.io>