Implement cluster reset command
`pxf cluster reset` will call local pxf-service script with 'reset' as a
command.

Both the local and remote versions prompt the user with a warning. For
the local side, this can be overridden with a `-f` or `--force`, and
that is how the cluster version of the command issues the remote
commands once the user has given permission.

The reset includes deletion of
`$PXF_HOME/{run,pxf-service,conf/pxf-private.classpath}` and resetting
the `$PXF_HOME/conf/pxf-env-default.sh` to an un-initialized condition.

To avoid loss of user configuration data, `pxf [cluster] reset` doesn't
touch `$PXF_CONF`.

Co-authored-by: Oliver Albertini <oalbertini@pivotal.io>
Co-authored-by: Raymond Yin <ryin@pivotal.io>