Docker pipeline: use Dockerfile that triggered build
Previously we were `get`ing the `pxf_src` and Dockerfile resources
separately, and using the Dockerfile from `pxf_src`. This allowed
situations where the Dockerfile being used in the task was not the one
that actually triggered the build.

With the approach used in this commit, the pipeline task always uses the
Dockerfile that triggers it. The downside is that if you want to
re-trigger the docker image manually (to re-cache stuff on the image)
you will either need to temporarily change the pipeline to use the
Dockerfile from `pxf_src`, or push some change to the Dockerfile.

Authored-by: Oliver Albertini <oalbertini@pivotal.io>