Create new test group "security"
This new test group includes tests that are meant to be run with
Kerberos.

Hbase, and some of the CloudAccessTests are not currently included. This
should be remedied later on.

This commit also includes splitting of storeAsOrcAllTypes into two
tests, one that runs on Hive versions 1 & 2, and another that only runs
on Hive 1, due to differences with timestamps in the two versions.

Co-authored-by: Oliver Albertini <oalbertini@pivotal.io>
Co-authored-by: Francisco Guerrero <aguerrero@pivotal.io>