Enable multinode testing against GCP dataproc
This commit enables use of GCP dataproc Hadoop clusters with and
without Kerberos and adds a pipeline job for testing against secure
dataproc.

Because Terraform and thus CCP doesn't support Kerberized dataproc, we
have to use our own scripts to create/destroy the Hadoop clusters.
Timing of the destruction of the cluster is controlled by the parameter
in pxf-multinode-params.yml `dataproc-destroy-timeout`.

There is a new KERBEROS (true/false) parameter for installing dataproc,
PXF and testing PXF multinode.

We also remove redundant code for installation of PXF in
test_pxf_multinode.bash.

Co-authored-by: Oliver Albertini <oalbertini@pivotal.io>
Co-authored-by: Francisco Guerrero <aguerrero@pivotal.io>