Automation: stop using ssh tunnels
Previously we used ssh tunnels to keep using the localhost when talking
to zookeeper and GPDB. This was problematic because when jobs failed, it
would take two extra hours before the pipeline would actually go red.

Also it's good to actually point the tests to another host when
not on localhost. We use COPY command on GPDB so files that will be used
in the COPY command must be scp'd over to GPDB master.

Authored-by: Oliver Albertini <oalbertini@pivotal.io>