A new UGI is created every time in some cases
When security is disabled, a new UserGroupInformation is created every
time when the pxf.service.kerberos.principal or
pxf.service.kerberos.keytab or
hadoop.kerberos.min.seconds.before.relogin properties are set in the
configuration. This can occur when the pxf-site.xml template is copied
into the server configuration of the non-secured server. Even though the
values in these properties are dummy values that take no effect, but
they were still being used by the LoginSession for non-secure servers.