Log the number of fragments retrieved just once
When the fragment cache is enabled, we are logging the number of
fragments returned by the fragmenter for every thread requesting the
fragment list. We observed in large clusters this was producing a large
amount of information being logged at the info level. With this
approach, only the thread performing the fragmentation call will log at
info level, while other threads should still log at debug level.

When the fragment cache is disabled, all threads will log this
information at info level.