Minor code optimizations (#288)
* Minor code optimizations

While profiling the code, I noticed some hotspots in

DataType#isTextForm: Optimized by using an array, iterating over a small
array is significantly faster than using a HashSet.

GreenplumCSV#toCsvField: I noticed a high amount of calls to
String.length. Assign the length to a variable once and use it to
optimize it.