SecureLoginTest: Fix test pollution
SecureLoginTest has been failing sporadically on centos7 in Concourse
during the compile_pxf job. It turns out that if one of these tests:

* org.apache.hadoop.security.LoginSessionTest
* org.apache.hadoop.security.PxfUserGroupInformationTest

runs before SecureLoginTest, we get an error when SecureLogin tries to
get the default realm when looking up the user. This has to do with
test pollution when newing up Hadoop `User`s, which end up
setting the default Kerberos realm for the JVM. This is the Hadoop code
that up does that[0].

This commit makes sure that

* if we change system properties, we change them back after the tests
* we have default realm set to something so that we don't get empty
  string

[0] https://github.com/apache/hadoop/blob/826afbeae31ca687bc2f8471dc841b66ed2c6704/hadoop-common-project/hadoop-auth/src/main/java/org/apache/hadoop/security/authentication/util/KerberosName.java#L86

Co-authored-by: Oliver Albertini <oalbertini@pivotal.io>
Co-authored-by: Alex Denissov <adenissov@pivotal.io>