Parquet: Right trim char types during insert
Currently, there is an issue when inserting parquet data with char types
when they are padded. For example,

    CREATE WRITABLE EXTERNAL TABLE foo (id int, code char(5))
    LOCATION ('pxf://bucket/path?PROFILE=s3:parquet&SERVER=s3')
    FORMAT 'CUSTOM' (formatter='pxfwritable_export');

    INSERT INTO foo VALUE (1, '95126'), (2, '9'), (3, '9');

This will cause postgres to right pad tuples 2 and 3 with whitespace
after the char '9', which will be sent to PXF Server as '9    ', and
we'll store it as such in a parquet file. However, this is inconsistent
with Hive's behavior and it is also causing issues with predicate
pushdown. Queries like the following will produce no tuples:

    SELECT * FROM foo WHERE code = '9';

With this patch, we right-trim values of type char coming from Greenplum
before storing them to parquet. This brings us on-par with the way Hive
stores char types and it fixes predicate pushdown.