Parquet: Transform char predicates with trailing whitespace
When a predicate for a char type has trailing whitespace we transform
the query to c1 = 'a  ' to (c1 = 'a  ' OR c1 = 'a'). This allows us to
push the original predicate, but also it provides consistency with the
way we are storing char types in parquet. When the operator is a not
equals the query is transformed from c1 <> 'a  ' to (c1 <> 'a  ' AND c1
<> 'a').

Renamed the SupportedParquetPrimitiveTypePruner class to
ParquetOperatorPrunerAndTransformer to take into account the new
functionality of the class.