Add User Option to allow invalid input paths (#336)
While running performance tests, some external tables did not have any
data, and when writting data to the partition, they did not create the
expected path while reading. To illustrate this better, this is a sample
code of what happened:

    CREATE TABLE foo (id int, some_date date)
    DISTRIBUTED BY (id)
    PARTITION BY RANGE (some_date)
    (start('1992-01-01') INCLUSIVE end ('1998-12-31') INCLUSIVE every
    (30), default partition others);

The stamement above will create 87 partitions. In our experiment, some
of the partitions did not have any data (there were no records for some
date ranges).

We then replace the internal partition with an external partition, as
illustrated below:

    CREATE EXTERNAL TABLE replace_partition_87 (id int, some_date date)
    LOCATION (
    'pxf://bucket/non-existent-path/?PROFILE=s3:parquet&SERVER=s3'
    ) FORMAT 'CUSTOM' (formatter = 'pxfwritable_import');

    ALTER TABLE foo
    EXCHANGE PARTITION FOR (RANK(87)) WITH TABLE replace_partition_87
    WITHOUT VALIDATION;

Because some of the external tables end up with no data, queries fail
when reading the original partitioned table `foo`. This commit
introduces a new user property `IGNORE_INVALID_INPUT`, which when set to
true, it will ignore the error and return an empty list of fragments.
The new table definition would look like this:

    CREATE EXTERNAL TABLE replace_partition_87(id int, some_date date)
    LOCATION (
    'pxf://bucket/non-existent-path/?PROFLE=s3:parquet&IGNORE_INVALID_INPUT=true'
    ) FORMAT 'CUSTOM' (formatter = 'pxfwritable_import');

This commit does not change the default behaviour of erroring out when
the partition is missing, but rather gives users an option to not error
out when the input path is missing.