pxf-dev-base/Dockerfile: stop masking error rc
A recent commit added an if statement to the bash in this Dockerfile,
but didn't add `&&` at the end. As a result, if anything failed before
that, the non-zero return code was lost there. The remaining statements
are if statements, so they terminate with rc = 0. We ended up with green
pipelines, but bad builds.

Authored-by: Oliver Albertini <oalbertini@vmware.com>