Hive: Support column projection on Hive Profiles (#333)
Currently, Hive and HiveRC profiles do not support column
projection. Add support for column projection for these profiles at the
PXF level. PXF can determine the following scenarios:

1. The number of columns of the file is higher than the number of
   columns defined at the Greenplum table definition.
2. The number of columns of the file is less than the number of columns
   defined at the Greenplum table definition.
3. A mixed number of columns is present in the files we are accessing.

For this to work, the column names on the Greenplum table definition
must match the column names on the Hive table definition, as the
projection is based on column names rather than column positions.

* Log warn when the column is not present in the file, but defined in greenplum

* Allow a subset of the Hive schema to be defined

We allow a subset of the Hive schema columns to be defined in the List
of ColumnDescriptors, but we validate that all the names in the
ColumnDescriptor list is present in the Hive schema. If not present we
error out

* Hive: Validate that ALTERED hive tables can still be read

Add tests that:
1. Create a hive table
2. Read hive table
3. Alter hive table (add column)
4. Insert new records in hive table
5. Read Greenplum external table accessing the hive table
6. ALTER Greenplum external table with new column from hive
7. Read Greenplum external table with new column from hive

* Remove Hive tests that are no longer used

- Remove toLowerCase checking because Hive metastore always lowercases its schema
- Add clarification notes about why we need both i, j and structIndex