Fix license-finder in the norsk pipeline
Recently we experienced issues with the norsk pipeline for PXF. The
issue is a failure when scanning the `pxf/automation` Maven project.

Within the license finder code, this command is run:

mvn help:evaluate -Dexpression=project.parent -q -D forceStdout

which aims to find the project parent. The correct response is:

null object or invalid expression

because there is no parent to automation/pom.xml.

However, the mvn command fails because mvn expects an environment
variable, PXF_TMP_LIB, to be set. mvn is normally only run from our
automation/Makefile, which sets PXF_TMP_LIB.  For automation it points
to a directory with symlinks to jars from $PXF_HOME/lib. Historically,
PXF_TMP_LIB=$PXF_HOME/tmp.

This commit changes that in favor of a hardcoded value in the user's
home directory: $HOME/pxf_tmp_lib. Symlinks will be created there
linking to jar files in $PXF_HOME/lib. This way mvn will not fail when
the above command is run, because HOME will be defined, and pointing to
an absolute path.

Authored-by: Oliver Albertini <oalbertini@vmware.com>