Add additional information regarding the pxf.service.user.name property
Add additional information regarding the pxf.service.user.name property
in pxf-site.xml. This property has been causing confusion in users and
it seems they are modifying the value of this property even in
environments where they should not. In this commit, I am moving the
property to the bottom of the file, and I add a comment in hopes that
users do read this comment, because they usually don't read the
<description> tag

Co-authored-by: Alexander Denissov <adenissov@pivotal.io>