Continuous Integration: independent release process
This commit introduces several new pipelines to aid a PXF release
process independent of any underlying SQL database. This commit thus
includes scripts for making and managing a PXF RPM, which includes the C
extension, whose code now lives under external-table/. These are the new
pipelines:

* pivnet_artifacts: a pipeline that fetches releases of GPDB from
  PivNet, aka TanzuNet
* pxf-build: pipeline that builds and releases new PXF artifacts
* dev-pxf-build: dev version of pxf-build, much abbreviated
* pxf-certification: pipeline that tests existing PXF artifacts against
  new versions of Greenplum
* pxf_pr_pipeline: new PR pipeline which uses the new build process

Version is now managed at the top level, in a file `version`. By
changing this file to a non-SNAPSHOT version, we can create a release
using the pxf-build pipeline. This causes a new tag/release to be
triggered on GitHub (Releases API), as well as a new SNAPSHOT version to
be committed.

Further, by uploading an OSL file to the appropriate cloud location,
we trigger publication of our release to the Release Engineering team.

Along with an independent release, we are now also installing PXF not
under GPHOME but separately in PXF_HOME. Thus this commit also has
changes to the PXF CLI to reflect the new install location. This
includes a new command called `pxf [cluster] register` to re-install the
PXF C extension under GPHOME.

Other noteworthy changes include:

* The new pipelines make use of smaller images that don't include PXF
  dependencies, but instead pull in those dependencies via tarballs that
  are pre-built.
* The new pattern for setting pipelines is to have targets in
  concourse/Makefile.
* Cloud tests have been hardened to reduce flakiness
* In CI scripts we source ~gpadmin/.pxfrc which is cached on each image
* Slack integrations have been added for pxf-certification, pxf-build,
  and dev-pxf-build pipelines

Co-authored-by: Oliver Albertini <oalbertini@vmware.com>
Co-authored-by: Alexander Denissov <adenissov@pivotal.io>
Co-authored-by: Francisco Guerrero <aguerrero@pivotal.io>