Add scalar complex types to HdfsWritableAvroTest
We convert it them to strings in Avro, but it's good to verify that it
works. Later we can support writing to the actual logical types[0] in
Avro.

* date
* time
* timestamp
* timestamp with time zone

Also we add testing for primitive character types:

* char
* varchar

These types become strings in Avro, but can still be read back into the
corresponding char types in Postgres.

[0] https://avro.apache.org/docs/current/spec.html#Logical+Types

Authored-by: Oliver Albertini <oalbertini@vmware.com>