Allow skipping the header for *:text profiles (#410)
We introduce a new parameter "SKIP_HEADER_COUNT", which when provided
for *:text profiles, it will skip the N first lines of the first split
of each file. For example, SKIP_HEADER_COUNT=2 will skip the first 2
lines of each remote file and will only stream back starting from the
3rd line. If SKIP_HEADER_COUNT is greater than the number of lines of a
file, no data will be streamed back to the client.