Hive: Report the correct error message from HiveMetaStoreClientCompatibility1xx (#415)
PXF is not reporting the causing error message when accessing Hive
version 1.x. PXF is also not reporting the causing error message during
initialization of HiveClient. For example, when a RuntimeException is
thrown when the HiveMetaStoreClientCompatibility1xx failed to
instantiate with a MetaException cause. The reported error is
java.lang.RuntimeException: Unable to instantiate
org.apache.hadoop.hive.metastore.HiveMetaStoreClientCompatibility1xx,
however the error above does not report the original cause of the
exception, making debugging harder for users/administrators.