Remove threadSafe locks
Currently we allow the bzip2 codec in when writing text and Avro formats
to an external source.

However, we do not run concurrent threads for bzip2 since it's not
considered to be thread-safe. This might have been fixed by now, so we
should create an automation test to read/write files using bzip2
compression in multi-threaded mode to have assurance that bzip2 is now
thread safe.

Earlier versions of CBZip2InputStream, the library used by hadoop to
decode bzip2 files was not thread-safe. This shortcoming has been
addressed in version 2.5.0 of hadoop [1]. Because PXF bundles newer
hadoop libraries, it doesn't need to lock anymore, as these libraries
are thread-safe now.

[1] https://issues.apache.org/jira/browse/HADOOP-10614