Hard-code git branch to 'master' when setting prod build pipeline
When setting the production build pipeline for PXF, the git branch
should always be set to master. Hard-coding it to 'master' safegaurds
against setting the pipeline when not on master. It also allows
state-checker to correctly check the build pipeline.

When the Conourse git resource clones, it will checkout a particular
commit causing the local working copy to be detached and the queried
branch name will be reported as `HEAD`.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>