Release UGI if there is an error during the filter execution
When the filter execution fails, for example when accessing a
non-existing HCFS path, the fragmenter call will throw an
InvalidInputException. This causes all segments to leave the UGI entry
available, and the FileSystem associated to the UGI does not get closed
until the UGI entry is expired 15 minutes earlier. This commit eagerly
closes UGI resources when an exception occurs.