Remove Configuration from SessionId
The Configuration in SessionId was only being used to check whether
security is enabled for the given server. But the Configuration being in
the SessionId brings a lot of issues, because when SessionIds are being
kept in the UGICache, they would retain a lot of associated objects in
memory. This would cause in some cases hundreds of megabytes in memory
just for the purpose of keeping a single field from Configuration. This
commit we only store the `isSecurityEnabled` boolean in the SessionId,
as it is the only piece of information that we need to keep.