Optimize hive metadata (#474)
* HiveUserData: Remove filterInFragmenter

The `filterInFragmenter` field is not being consumed by anyone. Removing
it.

* HiveUserData: Remove allColumnTypes from HiveUserData

* Parse HiveUserData once

Parse the HiveUserData object once during the accessor call. The parsed
HiveUserData is then shipped to the resolver using the
RequestContext#metadata field.

* HiveUserData: Remove serdeClassName from HiveUserData

The serde class name is available in the properties used for serde.

* HiveUserData: Remove inputFormatName from HiveUserData

The input format class name can be retrieved from the serde properties
by accessing the "file.inputformat" property.

* HiveUserData: Remove allColumnNames from HiveUserData

The same information is retrieved from the "columns" property of the
metadata properties that were retrieved during fragmentation.

* HiveUserData: Remove colTypes from HiveUserData

The column types can be retrieved from the metadata properties that are
shipped by the Fragmenter.

* HiveUserData: Remove skipHeader from HiveUserData

The skip header property is part of the metadata properties that we ship
from the fragmenter call.

* HiveUserData: Remove delimiter property from HiveUserData

Merge map where the property is coming from with the properties. Parse
delimiter at the Resolver only if needed.

* wip

* Remove HiveUserData

Only pass the schema properties from the fragmenter to the
accessor/resolver.

* Fix automation tests

* Hardcode HOSTS as it is no longer being used

* Remove properties not used

* Use Kryo to serialize properties

* Use a custom Properties serializer

* Add javadocs and tests

* Address PR feedback

* Remove delimiter code