Add PXF CLI test to multinode test (#479)
After the RPM work was merged, the CLI test pipeline stopped working.
This PR restores the CLI tests, but it adds the test as a task in the
multinode impersonation job. This way, it reuses the setup, which is
expensive, and then the automation tests run after the CLI tests are
successful.