Build the regression-tools tarball (#510)
PXF regression tests require some tools, for example the avro tools jar,
to perform data preparation. We build the regression-tools tarball and add all
the jars needed for data preparation, we then push it to the
dependencies bucket. This tarball can then be later consumed by the
regression test suite.