Update PXF build pipeline for PXF 6+ releases
PXF 6+ drops support for RHEL6 in both GP5 and GP6. In previous PXF
releases, a binary tarball for PXF compatible with GP5 on RHEL6 was
produced in the build pipeline. The release process now builds a binary
tarball of PXF for GP5 on REHL7. To better help distinguish the tarball
by GP major version and platform, this information is included in the
artifact filename.

With the changes in the artifact filenames and the set of released
artifacts, the template email to releng is updated as well.

The PXF release process currently uploads ODP and OSL artifacts to a
GP6-specific location in blob store and directs the RelEng team to use
the same artifacts for PXF-GP5 releases. The RelEng team would prefer
that the ODP and OSL artifacts get uploaded to GP5 and GP6 locations.

Instead of adding additional template variables for GP5 and GP6 paths
for ODP and OSL, the existing template variables are updated to remove
the GP major version from the path. This is appended in the CI pipeline
instead. Since all release artifacts for RelEng are stored in the same
bucket, the template variables for the bucket name are simplified as
well.

PXF ODP and OSL artifacts for both GP5 and GP6 are now stored in Google
Cloud Storage so there isn't much benefit to parameterizing the protocol
(i.e., `s3` vs `gs`).

The normal flow for GP5 and GP6 RPMs is to fetch from pivnet and store
in a UD-specific bucket. Since GP7 isn't available via pivnet yet, we
use the location that gpdb_master publishes release candidate RPMs to.

[skip ci]

Co-authored-by: Ashuka Xue <axue@vmware.com>
Co-authored-by: Bradford D. Boyle <bradfordb@vmware.com>