Update automation to run on either JDK8 or JDK11 (#643)
* automation: update Hadoop dependencies to match w/ server

Adding support for JDK11 requires a newer Hadoop version.

Co-authored-by: Bradford D. Boyle <bradfordb@vmware.com>
Co-authored-by: Ashuka Xue <axue@vmware.com>

* automation: set TERM environment variable when executing remote commands

The way automation runs test is that it connects via SSH to the
localhost, spawning an instance of your login shell (e.g., `bash` or
`zsh`). Somehow, the `TERM` environment variable gets set to or defaults
to `xterm-256color`, which results in extraneous ANSI escape codes
printing to standard out. Since automation reads standard out to parse
the return codes of commands it executes, these extra escape codes cause
it to incorrectly interpret the output. To resolve this issue,
automation explicitly set the value of `TERM` when running commands. The
`dumb` terminal capabilities assumes that simplest set of terminal
capabilities and appears to be a safe default.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>

* automation: Document required setup on newer Linux distros

- document required SSH setup on localhost including RSA key and
  KexAlgorithms
- document installing python2-pip and tinc dependencies cleanly

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>

* automation: make an empty jar

In JDK11, creating an zero-length file (with `touch`) causes an error as
the file is not a valid jar/zip file.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>

Co-authored-by: Ashuka Xue <axue@vmware.com>