Remove the error-causing check (#662)
This check is not necessary. executeUpdate() throws an exception in case of an error and doesn't return 0; it is also contrary to the SQL semantics for an INSERT statement to return anything other than 1 when no error happens. That is, checking the return value is redundant. Also, the implementation of the function in hive jdbc driver always returns 0. Because of this, even in the case of a successful INSERT, an error was thrown.