Add CI tasks and scripts for creating Hadoop cluster with IPA (#712)
The Terraform module in concourse/terraform/ipa-multinode-hadoop
provisions infrastructure for the cluster in Google Cloud. The module
creates a VM for running an instance of IPA server to provide a Kerberos
key distribution center (KDC) that supports constrained delegation. VMs
for running NamdeNode servers and DataNode servers are also created.

To avoid leaking infrastructure from failed CI jobs, a new Concourse
pipeline called hadoop-cluster-cleaner is included. This pipeline uses
the ccp_reaper utility from gp-concourse-cluster-provisioner to find and
destroy infrastructure that is no longer in use. The reaper utility
needs access to the Terraform module that provisioned the infrastructure
so it is included here.

The Ansible play in concourse/ansible/ipa-multinode-hadoop will
configure IPA and Hadoop on the VMs created with the Terraform module.
The play installs IPA server on the dedicated VM and then installs
Hadoop (using singlecluster archive) on NameNode and DataNode VMs. The
Hadoop nodes are also configured as IPA clients. A Concourse task for
consuming the metadata produced by the Terraform Concourse resource and
running ansible-playbook is included.

The existing multinode CI tasks and scripts are updated to setup PXF to
use the new Hadoop cluster if the cluster environment files are
provided. The build pipeline does not currently create the cluster so
these setup steps are skipped.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>