ci: include PXF major version in cloud storage path (#801)
The build pipelines for PXF 5.16 and 6 use the same path for storing
snapshot tarballs. The pipelines can distinguish artifacts produced by
the different pipelines by correlating the files generation with the CI
job that produced the artifact. CI jobs outside of the build pipelines
cannot distinguish between snapshot builds of PXF 5.16 and 6 and will
randomly get one or the other depending on which ran last.

This commit moves PXF snapshot tarballs to be stored in a location that
includes the PXF major version. This will allow downstream consumers of
snapshot builds to select which PXF major version to pull.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>