Add data buffer boundary checks to PXF extension (#814)
If a GPDBWritable message is malformed, it can cause PXF's custom
formatter to attempt a read outside of the valid boundaries of the
passed in data buffer. This results in a SIGSEGV that panics the segment
and leads to a failover.

This commit adds some basic boundary checks before reading values out of
the received data and reports an error if the length would exceed the
valid range of indices for the passed in data buffer. Additionally, this
commit updates the formatter code to check that the received column
types are consistent with the table's column descriptors for every
received tuple. Previously, this check was only done for the first
received tuple. Finally, this commit also checks that the actual length
of every received string matches the length given in the GPDBWritable
message.

Error are logged at the FATAL level, which will keep the backend around
for a short amount of time so that someone can capture a core dump with
either gcore or gdp. The amount of time to keep the process around is
configured via gp_debug_linger.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>