Add write support for one-dimensional ORC arrays (#836)
This commit adds support for writing one-dimensional arrays of all currently supported primitive types into ORC files with a PXF-generated ORC schema. In the case of TIME and UUID array types, these are written to the ORC file as array<string> as ORC does not have TIME or UUID types. 

This commit also adds READ functionality for the date, timestamp and timestamp with timezone array types with ORC.